requirements.txt
asgiref
astroid
Django
django-clear-cache
django-cors-headers
django-rest-framework-mongoengine
djangorestframework
djongo
freeze
isort
lazy-object-proxy
mccabe
mongoengine
Pillow
pip
pylint
pylint-django
pylint-plugin-utils
pymongo
pytz
setuptools
six
sqlparse
toml
wheel
wrapt
